Clean up migration filter registration
